requirements.txt
# This file is @generated by PDM.
# Please do not edit it manually.

addict==2.4.0
aiohappyeyeballs==2.6.1
aiohttp==3.11.14
aiosignal==1.3.2
annotated-types==0.7.0
attrs==25.3.0
blinker==1.9.0
ccimport==0.4.4
certifi==2025.1.31
charset-normalizer==3.4.1
click==8.1.8
cmake==3.31.6; platform_system == "Linux" and platform_machine == "x86_64"
colorama==0.4.6; platform_system == "Windows"
configargparse==1.7
contourpy==1.3.1
cramjam==2.9.1
cumm-cu120==0.4.11
cycler==0.12.1
dash==3.0.1
docker-pycreds==0.4.0
fastjsonschema==2.21.1
fastparquet==2024.11.0
filelock==3.18.0
fire==0.7.0
flask==3.0.3
fonttools==4.56.0
frozenlist==1.5.0
fsspec==2025.3.0
gitdb==4.0.12
gitpython==3.1.44
idna==3.10
importlib-metadata==8.6.1
isort==6.0.1
itsdangerous==2.2.0
jinja2==3.1.6
joblib==1.4.2
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
jupyter-core==5.7.2
kaleido==0.2.1
kiwisolver==1.4.8
lark==1.2.2
lit==18.1.8; platform_system == "Linux" and platform_machine == "x86_64"
llvmlite==0.44.0
markupsafe==3.0.2
matplotlib==3.10.1
mpmath==1.3.0
multidict==6.2.0
narwhals==1.32.0
nbformat==5.10.4
nest-asyncio==1.6.0
networkx==3.4.2
ninja==1.11.1.4
numba==0.61.0
numpy==1.26.4
nvidia-cublas-cu11==11.10.3.66; platform_system == "Linux" and platform_machine == "x86_64"
nvidia-cuda-cupti-cu11==11.7.101; platform_system == "Linux" and platform_machine == "x86_64"
nvidia-cuda-nvrtc-cu11==11.7.99; platform_system == "Linux" and platform_machine == "x86_64"
nvidia-cuda-runtime-cu11==11.7.99; platform_system == "Linux" and platform_machine == "x86_64"
nvidia-cudnn-cu11==8.5.0.96; platform_system == "Linux" and platform_machine == "x86_64"
nvidia-cufft-cu11==10.9.0.58; platform_system == "Linux" and platform_machine == "x86_64"
nvidia-curand-cu11==10.2.10.91; platform_system == "Linux" and platform_machine == "x86_64"
nvidia-cusolver-cu11==11.4.0.1; platform_system == "Linux" and platform_machine == "x86_64"
nvidia-cusparse-cu11==11.7.4.91; platform_system == "Linux" and platform_machine == "x86_64"
nvidia-nccl-cu11==2.14.3; platform_system == "Linux" and platform_machine == "x86_64"
nvidia-nvtx-cu11==11.7.91; platform_system == "Linux" and platform_machine == "x86_64"
open3d==0.19.0
packaging==24.2
pandas==2.2.3
pccm==0.4.16
pillow==11.1.0
platformdirs==4.3.7
plotly==6.0.1
portalocker==3.1.1
propcache==0.3.1
protobuf==5.29.4; python_version > "3.9" or sys_platform != "linux"
psutil==7.0.0
pybind11==2.13.6
pydantic==2.10.6
pydantic-core==2.27.2
pyparsing==3.2.3
pyquaternion==0.9.9
python-dateutil==2.9.0.post0
pytorch-metric-learning==2.8.1
pytz==2025.2
pywin32==310; sys_platform == "win32" and platform_python_implementation != "PyPy" or platform_system == "Windows"
pyyaml==6.0.2
referencing==0.36.2
requests==2.32.3
retrying==1.3.4
rpds-py==0.24.0
scikit-learn==1.6.1
scipy==1.15.2
sentry-sdk==2.24.1
setproctitle==1.3.5
setuptools==78.0.2
six==1.17.0
smmap==5.0.2
spconv-cu120==2.3.6
strictyaml==1.7.3
sympy==1.13.3
tabulate==0.9.0
tailwind-colors==1.3.0
termcolor==2.5.0
threadpoolctl==3.6.0
torch==2.0.1
torch-cluster @ https://data.pyg.org/whl/torch-2.0.0%2Bcu117/torch_cluster-1.6.3%2Bpt20cu117-cp311-cp311-linux_x86_64.whl
torch-geometric==2.6.1
torch-scatter @ https://data.pyg.org/whl/torch-2.0.0%2Bcu117/torch_scatter-2.1.2%2Bpt20cu117-cp311-cp311-linux_x86_64.whl
torch-sparse @ https://data.pyg.org/whl/torch-2.0.0%2Bcu117/torch_sparse-0.6.18%2Bpt20cu117-cp311-cp311-linux_x86_64.whl
tqdm==4.67.1
traitlets==5.14.3
triton==2.0.0; platform_system == "Linux" and platform_machine == "x86_64"
typing-extensions==4.12.2
tzdata==2025.2
urllib3==2.3.0
wandb==0.19.8
werkzeug==3.0.6
wheel==0.45.1; platform_system == "Linux" and platform_machine == "x86_64"
yarl==1.18.3
zipp==3.21.0

